backend/requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
psycopg2==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
tzdata==2025.2
